capitulo_8/recursos/ejercicio_1/requirements.txt
attrs==19.3.0
cachetools==4.0.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
clickclick==1.2.2
connexion==2.4.0
Flask==1.1.1
google-api-core==1.16.0
google-auth==1.11.2
google-cloud-core==1.3.0
google-cloud-storage==1.26.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
idna==2.8
importlib-metadata==0.23
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.14.1
jsonschema==3.1.1
MarkupSafe==1.1.1
more-itertools==7.2.0
numpy==1.18.1
openapi-spec-validator==0.2.8
pandas==1.0.1
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo==3.9.0
pyrsistent==0.15.5
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
rsa==4.0
scikit-learn==0.22.1
scipy==1.4.1
six==1.13.0
sklearn==0.0
swagger-ui-bundle==0.0.6
urllib3==1.25.7
Werkzeug==0.16.0
zipp==0.6.0
